# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CareSync Telemedicine System API
  description: Production-grade API specification for Amrutam's CareSync Telemedicine System.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /doctors:
    get:
      summary: Search and filter verified doctors
      parameters:
        - name: specialization
          in: query
          required: false
          schema:
            type: string
        - name: feeMin
          in: query
          required: false
          schema:
            type: number
        - name: feeMax
          in: query
          required: false
          schema:
            type: number
        - name: experienceMin
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Array of verified doctors.
...
  /consultations:
    post:
      summary: Book an appointment slot (Patient Only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - slotId
              properties:
                slotId:
                  type: string
                  format: uuid
                  example: "68cc6b80-48a9-4e16-b030-ed5948c10452"
                reasonForVisit:
                  type: string
                  example: "Mild fever and sore throat."
      responses:
        '201':
          description: Slot locked and appointment scheduled.
        '409':
          description: Slot has already been booked by another transaction.
    get:
      summary: List consultations for active session (Patient/Doctor)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of consultations.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken